requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==2.1.1
click==8.1.3
docker-pycreds==0.4.0
filelock==3.9.0
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.10
GitPython==3.1.30
huggingface-hub==0.12.0
hydra-core==1.3.1
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
joblib==1.2.0
lightning-utilities==0.6.0.post0
multidict==6.0.4
numpy==1.21.6
omegaconf==2.3.0
packaging==23.0
pandas==1.3.5
pathtools==0.1.2
Pillow==9.4.0
protobuf==4.21.12
psutil==5.9.4
python-dateutil==2.8.2
pytorch-lightning==1.9.1
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.15.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchmetrics==0.11.1
torchvision==0.14.1+cu116
tqdm==4.64.1
transformers==4.26.1
typing_extensions==4.4.0
urllib3==1.26.14
wandb==0.13.10
yarl==1.8.2
zipp==3.13.0
